moved Header and BackToHome components to separate folder
