+++ src/routes/LandingPage.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Header from "../components/Header";
+import Header from "../components/shared/Header";
 
 function LandingPage() {
   return (
