+++ src/routes/host/QuizInProgress.tsx
@@ -1,6 +1,6 @@
 import { useParams } from "react-router-dom";
-import Header from "../../components/Header";
-import BackToHome from "../../components/BackToHome";
+import Header from "../../components/shared/Header";
+import BackToHome from "../../components/shared/BackToHome";
 
 export default function QuizInProgress() {
   const { room_id: roomId } = useParams();
